Added Default LostItem Image, implemented Lost and Found Expiration functionality.

+++ Sunridge/Pages/Admin/LostAndFound/Upsert.cshtml.cs
@@ -68,18 +68,24 @@ namespace Sunridge.Pages.Admin.LostItem
             #region Image Upload
             if (LostItemObj.LostItem.Id == 0)
             {
-                // Upload and save image
-                string fileName = Guid.NewGuid().ToString();
-                var uploads = Path.Combine(webRootPath, @"images\lostItems");
-                var extension = Path.GetExtension(files[0].FileName);
-
-                using (var fileStream = new FileStream(Path.Combine(uploads, fileName + extension), FileMode.Create))
+                if (LostItemObj.LostItem.Image == null)
                 {
-                    files[0].CopyTo(fileStream);
+                    LostItemObj.LostItem.Image = "default.png";
                 }
+                else
+                {
+                    // Upload and save image
+                    string fileName = Guid.NewGuid().ToString();
+                    var uploads = Path.Combine(webRootPath, @"images\lostItems");
+                    var extension = Path.GetExtension(files[0].FileName);
 
-                LostItemObj.LostItem.Image = fileName + extension;
+                    using (var fileStream = new FileStream(Path.Combine(uploads, fileName + extension), FileMode.Create))
+                    {
+                        files[0].CopyTo(fileStream);
+                    }
 
+                    LostItemObj.LostItem.Image = fileName + extension;
+                }
                 _unitOfWork.LostItem.Add(LostItemObj.LostItem);
             }
             #endregion
@@ -99,10 +105,12 @@ namespace Sunridge.Pages.Admin.LostItem
                     var extension = Path.GetExtension(files[0].FileName);
 
                     var filePath = Path.Combine(webRootPath, objFromDb.Image.TrimStart('\\'));
-
-                    if (System.IO.File.Exists(filePath))
+                    if (LostItemObj.LostItem.Image != "default.png")
                     {
-                        System.IO.File.Delete(filePath);
+                        if (System.IO.File.Exists(filePath))
+                        {
+                            System.IO.File.Delete(filePath);
+                        }
                     }
 
                     using (var fileStream =
